general issues solved like xml comment removed from repository classes and code refactoring
