+++ src/Promact.TestCaseManagement.Repository/ApplicationClass/ProjectAC.cs
@@ -1,4 +1,4 @@
-﻿namespace Promact.TestCaseManagement.Repository.ApplicationClass.External
+﻿namespace Promact.TestCaseManagement.Repository.ApplicationClass
 {
     public class ProjectAC
     {
