+++ src/Promact.TestCaseManagement.Repository/ApplicationClass/UserDetailWithProject.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Generic;
 
-namespace Promact.TestCaseManagement.Repository.ApplicationClass.External
+namespace Promact.TestCaseManagement.Repository.ApplicationClass
 {
     public class UserDetailWithProject
     {
         public UserAC UserAc { get; set; }
 
-        public List<ProjectAC> ListOfProject { get; set; }
+        public IEnumerable<ProjectAC> ListOfProject { get; set; }
     }
 }
